Fix helicopter edge-flip: wrap Abot to u8 in draw_helicoper_part
Abot modelled the Z80 A register after ADD A,B, which wraps mod 256.
Declared int, it didn't wrap, so values near a screen edge could reach
up to 509 before reaching draw_object_left_width_entrypt, which
branches on Awidth_bytes assuming an 8-bit wrapped value — sending the
sprite down the wrong path and drawing it on the opposite edge.

+++ C/libraries/ChaseHQ/Engine/Main.c
@@ -9249,7 +9249,7 @@ static void draw_helicoper_part(chqstate_t                *state,
   const bitmap_t *HLbitmap;   /* was HL */
   int             Bwidth;     /* was B */
   int             Atop;       /* was A */
-  int             Abot;       /* was A */
+  u8              Abot;       /* was A; Conv: u8 so += wraps mod 256 like the Z80 ADD A,B */
   int             C;          /* was C */
 
   BC_helipos = state->dhl_helipos;
